WIP: Send the Client-Secret to Joplin Server when authenticating with SAML
